Fix comprehensive golangci-lint type errors for O-RAN dashboard package
- Add missing type definitions for E2AP protocol, simulator, and SMO integration
- Resolve E2NodeConfigurationUpdate, SimulatedSubscription, RMRMessage types
- Fix E2SMKPMMetrics, NETCONFClient, and performance optimization types
- Remove duplicate testVulnerabilityProtection method in security tests
- Add comprehensive A1 Mediator, policy management, and service model types
- Fix field name mismatches and interface implementations
- Ensure dashboard package compiles successfully with all dependencies

+++ pkg/dashboard/advanced_smo_performance_optimizer.go
@@ -373,7 +373,7 @@ func (aspo *AdvancedSMOPerformanceOptimizer) OptimizeForThroughput(targetIPS int
 	}
 
 	// Update configuration
-	aspo.config.TargetThroughputIPS = targetIPS
+	aspo.config.TargetThroughputIPS = float64(targetIPS)
 
 	return nil
 }
